Fixed all compiler warnings

+++ source/include_local/macro_util.h
@@ -9,8 +9,8 @@ extern "C" {
 #include <stdbool.h>
 
 void IoTLib_perform_managed_array_bounds_check(const size_t managedArrayLength, const size_t managedArrayCapacity);
-void IoTLib_managed_key_value_array_getFunc_validate_key_index(const size_t keyIndex);
-bool IoTLib_managed_key_value_array_index_is_valid_index(const size_t keyIndex);
+void IoTLib_managed_key_value_array_getFunc_validate_key_index(const int keyIndex);
+bool IoTLib_managed_key_value_array_index_is_valid_index(const int keyIndex);
 
 #ifdef __cplusplus
 }
