+++ source/include_local/sensorID.h
@@ -8,7 +8,7 @@ extern "C" {
 #include <stddef.h>
 #include "typedefs.h"
 
-size_t IoTLib_sensorid_get_key_index(const IoTLib_SensorID* keyArray, const size_t arrayLength, const IoTLib_SensorID keyToSearchFor);
+int IoTLib_sensorid_get_key_index(const IoTLib_SensorID* keyArray, const size_t arrayLength, const IoTLib_SensorID keyToSearchFor);
 
 #ifdef __cplusplus
 }
